backend/requirements.txt
django
djangorestframework
openai
pydantic_ai
django-cors-headers